Rename utils to extensions folder
